pyproject.toml
[tool.poetry]
name = "evraz-gpt-unami"
version = "0.0.1"
description = "This project is solution for Hackathon Evraz"
authors = ["Mingchen Zhuge <mingchen.zhuge@kaust.edu.sa>", ""]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
tiktoken = "^0.8.0"
rich = "^13.9.2"
litellm = "^1.50.0"
tenacity = "^9.0.0"
numpy = "<2.0"
networkx = "^3.3"
spacy = "<3.8.0"
rank-bm25 = "^0.2.2"
sentence-transformers = "^3.1.1"
pandas = "^2.2.3"
docx = "^0.2.4"
markdown = "^3.7"
python-docx = "^1.1.2"
pypdf2 = "^3.0.1"
openpyxl = "^3.1.5"
python-pptx = "^1.0.2"
opencv-python = "^4.10.0.84"
beautifulsoup4 = "^4.12.3"
pylatexenc = "^2.10"
matplotlib = "^3.9.2"
tree-sitter-languages = "1.8.0"
grep-ast = "^0.3.3"
rapidfuzz = "^3.10.0"
tqdm = "^4.66.5"
logging = "^0.4.9.6"
tree-sitter = "0.21.3"
pytest = "^8.3.3"
tf-keras = "^2.17.0"
aiogram = "3.11.0"
telebot = ">=0.0.5,<0.1.0"
loguru = "^0.7.2"
pydantic-settings = ">=2.3.4,<2.4.0"
pyarrow = "^18.1.0"
dill = "^0.3.9"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
